requirements.txt
adbutils
airtest
allure-pytest
allure-python-commons
altgraph
apipkg
apkutils2
Appium-Python-Client
argon2-cffi
asgiref
astor
async-generator
atomicwrites
attrdict
attrs
Babel
backcall
bce-python-sdk
bleach
blinker
Brotli
cached-property
cachetools
certifi
cffi
cfgv
chardet
cigam
click
colorama
colorlog
comtypes
ConfigArgParse
coreapi
coreschema
cryptography
cssselect
cssutils
cycler
Cython
decorator
defusedxml
Deprecated
deprecation
dill
distlib
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-jwt
djangorestframework-yaml
drf-yasg
entrypoints
et-xmlfile
execnet
facebook-wda
Faker
filelock
filetype
flake8
Flask
Flask-Babel
Flask-BasicAuth
future
gevent
geventhttpclient
greenlet
h11
h2
h5py
har2case
hpack
hrpc
HttpRunner
hyperframe
identify
idna
imageio
imgaug
importlib-metadata
inflection
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
itypes
jdcal
jedi
jieba
Jinja2
joblib
jsonpath
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyterlab-pygments
jupyterlab-widgets
kaitaistruct
kiwisolver
ldap3
lmdb
locust
loguru
logzero
lxml
Markdown
MarkupSafe
matplotlib
mccabe
mistune
mitmproxy
more-itertools
msgpack
mss
multiprocess
mysql
mysqlclient
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nodeenv
notebook
numpy
opencv-contrib-python
opencv-python
openpyxl
packaging
pactverify
paddlenlp
paddleocr
paddlepaddle
pandas
pandocfilters
parso
passlib
pefile
pickleshare
Pillow
platformdirs
pluggy
pocoui
pre-commit
premailer
progress
prometheus-client
prompt-toolkit
protobuf
psutil
publicsuffix2
py
pyasn1
pyclipper
pycodestyle
pycparser
pycryptodome
pydantic
pydivert
pyelftools
pyflakes
Pygments
pyinstaller
pyinstaller-hooks-contrib
PyJWT
PyMySQL
pyOpenSSL
pyparsing
pyperclip
pyrsistent
pytest
pytest-forked
pytest-html
pytest-metadata
pytest-rerunfailures
pytest-xdist
python-dateutil
pytz
PyWavelets
pywin32
pywin32-ctypes
pywinauto
pywinpty
PyYAML
pyzmq
qtconsole
QtPy
requests
requests-toolbelt
retry
ruamel.yaml
ruamel.yaml.clib
scikit-image
scikit-learn
scipy
selenium
Send2Trash
seqeval
Shapely
shellcheck-py
six
sortedcontainers
sqlparse
terminado
testpath
text-unidecode
threadpoolctl
tifffile
toml
tornado
tqdm
traitlets
typing-extensions
uiautomator2
unittestreport
uritemplate
urllib3
urwid
virtualenv
visualdl
wcwidth
webencodings
websocket-client
weditor
Werkzeug
whichcraft
widgetsnbextension
win32-setctime
wrapt
wsproto
XlsxWriter
xmltodict
zipp
zope.event
zope.interface
zstandard
